+++ chia/cmds/beta_funcs.py
@@ -5,7 +5,8 @@ import sys
 from pathlib import Path
 from typing import Any, Callable, Optional
 
-from chia.cmds.cmds_util import format_bytes, prompt_yes_no, validate_directory_writable
+from chia.cmds.cmds_util import format_bytes, validate_directory_writable
+from chia.cmds.passphrase_funcs import prompt_yes_no
 from chia.util.beta_metrics import metrics_log_interval_max, metrics_log_interval_min
 from chia.util.chia_logging import get_beta_logging_config
 from chia.util.errors import InvalidPathError
