+++ chia/cmds/cmds_util.py
@@ -478,16 +478,6 @@ def format_minutes(minutes: int) -> str:
     return "Unknown"
 
 
-def prompt_yes_no(prompt: str) -> bool:
-    while True:
-        response = str(input(prompt + " (y/n): ")).lower().strip()
-        ch = response[:1]
-        if ch == "y":
-            return True
-        elif ch == "n":
-            return False
-
-
 def validate_directory_writable(path: Path) -> None:
     write_test_path = path / ".write_test"
     try:
